LayoutTests/imported/w3c/web-platform-tests/css/cssom/variable-names-expected.txt


PASS custom property '--a'
FAIL custom property '--a;b' assert_equals: appears on specified style (after serialization/re-parsing) expected 1 but got 0
PASS custom property '---'
FAIL custom property '--\' assert_equals: appears on specified style (after serialization/re-parsing) expected 1 but got 0
PASS custom property '--ab'
PASS custom property '--0'
